requirements.txt
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
deprecation==2.1.0
dotenv-python==0.0.1
grpcio==1.64.1
idna==3.7
llvmlite==0.43.0
lxml==5.2.2
numba==0.60.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pandas-datareader==0.10.0
protobuf==4.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scipy==1.13.1
six==1.16.0
ta==0.11.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b100
tzdata==2024.1
urllib3==2.2.2
